ファイル分割・型整理
